Refactor runtime tests into separate files and make tests more specific
