+++ runtime/src/info.rs
@@ -1,14 +1,8 @@
 // Licensed under the Apache-2.0 license
 
-use crate::{handoff::RtHandoff, Drivers, MAX_CERT_CHAIN_SIZE, PL0_PAUSER_FLAG};
-use arrayvec::ArrayVec;
-use caliptra_common::mailbox_api::{
-    FwInfoResp, GetIdevCertReq, GetIdevCertResp, GetIdevInfoResp, MailboxResp, MailboxRespHeader,
-    PopulateIdevCertReq,
-};
-use caliptra_drivers::{CaliptraError, CaliptraResult};
-use caliptra_x509::{Ecdsa384CertBuilder, Ecdsa384Signature};
-use zerocopy::{AsBytes, FromBytes};
+use crate::{handoff::RtHandoff, Drivers};
+use caliptra_common::mailbox_api::{FwInfoResp, GetIdevInfoResp, MailboxResp, MailboxRespHeader};
+use caliptra_drivers::CaliptraResult;
 
 pub struct FwInfoCmd;
 impl FwInfoCmd {
@@ -46,76 +40,5 @@ impl IDevIdInfoCmd {
             idev_pub_x: pub_key.x.into(),
             idev_pub_y: pub_key.y.into(),
         }))
-    }
-}
-
-pub struct IDevIdCertCmd;
-impl IDevIdCertCmd {
-    pub(crate) fn execute(cmd_args: &[u8]) -> CaliptraResult<MailboxResp> {
-        if cmd_args.len() <= core::mem::size_of::<GetIdevCertReq>() {
-            let mut cmd = GetIdevCertReq::default();
-            cmd.as_bytes_mut()[..cmd_args.len()].copy_from_slice(cmd_args);
-
-            // Validate tbs
-            if cmd.tbs_size as usize > cmd.tbs.len() {
-                return Err(CaliptraError::RUNTIME_MAILBOX_INVALID_PARAMS);
-            }
-
-            let sig = Ecdsa384Signature {
-                r: cmd.signature_r,
-                s: cmd.signature_s,
-            };
-
-            let Some(builder) = Ecdsa384CertBuilder::new(&cmd.tbs[..cmd.tbs_size as usize], &sig) else {
-                return Err(CaliptraError::RUNTIME_GET_DEVID_CERT_FAILED);
-            };
-
-            let mut cert = [0; GetIdevCertResp::DATA_MAX_SIZE];
-            let Some(cert_size) = builder.build(&mut cert) else {
-                return Err(CaliptraError::RUNTIME_GET_DEVID_CERT_FAILED);
-            };
-
-            Ok(MailboxResp::GetIdevCert(GetIdevCertResp {
-                hdr: MailboxRespHeader::default(),
-                cert_size: cert_size as u32,
-                cert,
-            }))
-        } else {
-            Err(CaliptraError::RUNTIME_INSUFFICIENT_MEMORY)
-        }
-    }
-}
-
-pub struct PopulateIDevIdCertCmd;
-impl PopulateIDevIdCertCmd {
-    pub(crate) fn execute(drivers: &mut Drivers, cmd_args: &[u8]) -> CaliptraResult<MailboxResp> {
-        if cmd_args.len() <= core::mem::size_of::<PopulateIdevCertReq>() {
-            let mut cmd = PopulateIdevCertReq::default();
-            cmd.as_bytes_mut()[..cmd_args.len()].copy_from_slice(cmd_args);
-
-            let cert_size = cmd.cert_size as usize;
-            if cert_size > cmd.cert.len() {
-                return Err(CaliptraError::RUNTIME_MAILBOX_INVALID_PARAMS);
-            }
-
-            let flags = drivers.persistent_data.get().manifest1.header.flags;
-            // PL1 cannot call this mailbox command
-            if flags & PL0_PAUSER_FLAG == 0 {
-                return Err(CaliptraError::RUNTIME_INCORRECT_PAUSER_PRIVILEGE_LEVEL);
-            }
-
-            let mut tmp_chain = ArrayVec::<u8, MAX_CERT_CHAIN_SIZE>::new();
-            tmp_chain
-                .try_extend_from_slice(&cmd.cert[..cert_size])
-                .map_err(|_| CaliptraError::RUNTIME_IDEV_CERT_POPULATION_FAILED)?;
-            tmp_chain
-                .try_extend_from_slice(drivers.cert_chain.as_slice())
-                .map_err(|_| CaliptraError::RUNTIME_IDEV_CERT_POPULATION_FAILED)?;
-            drivers.cert_chain = tmp_chain;
-
-            Ok(MailboxResp::default())
-        } else {
-            Err(CaliptraError::RUNTIME_INSUFFICIENT_MEMORY)
-        }
     }
 }
