+++ runtime/tests/runtime_integration_tests/main.rs
@@ -1,7 +1,16 @@
 // Licensed under the Apache-2.0 license
 
 mod common;
-mod ecdsa;
-mod hmac;
-mod integration_tests;
+mod test_boot;
+mod test_certs;
+mod test_disable;
+mod test_ecdsa;
+mod test_fips;
+mod test_hmac;
+mod test_info;
+mod test_invoke_dpe;
+mod test_mailbox;
 mod test_panic_missing;
+mod test_populate_idev;
+mod test_stash_measurement;
+mod test_tagging;
