+++ runtime/tests/runtime_integration_tests/test_fips.rs
@@ -1,0 +1,71 @@
+// Licensed under the Apache-2.0 license.
+
+use crate::common::run_rt_test;
+use caliptra_common::mailbox_api::{
+    CommandId, FipsVersionResp, MailboxReqHeader, MailboxRespHeader,
+};
+use caliptra_error::CaliptraError;
+use caliptra_hw_model::{HwModel, ModelError};
+use caliptra_runtime::FipsVersionCmd;
+use zerocopy::{AsBytes, FromBytes};
+
+#[test]
+fn test_fips_version() {
+    let mut model = run_rt_test(None, None, None);
+
+    model.step_until(|m| m.soc_mbox().status().read().mbox_fsm_ps().mbox_idle());
+
+    // VERSION
+    let payload = MailboxReqHeader {
+        chksum: caliptra_common::checksum::calc_checksum(u32::from(CommandId::VERSION), &[]),
+    };
+
+    let fips_version_resp = model
+        .mailbox_execute(u32::from(CommandId::VERSION), payload.as_bytes())
+        .unwrap()
+        .unwrap();
+
+    // Check command size
+    let fips_version_bytes: &[u8] = fips_version_resp.as_bytes();
+
+    // Check values against expected.
+    let fips_version = FipsVersionResp::read_from(fips_version_bytes).unwrap();
+    assert!(caliptra_common::checksum::verify_checksum(
+        fips_version.hdr.chksum,
+        0x0,
+        &fips_version.as_bytes()[core::mem::size_of_val(&fips_version.hdr.chksum)..],
+    ));
+    assert_eq!(
+        fips_version.hdr.fips_status,
+        MailboxRespHeader::FIPS_STATUS_APPROVED
+    );
+    assert_eq!(fips_version.mode, FipsVersionCmd::MODE);
+    assert_eq!(fips_version.fips_rev, [0x01, 0xaaaaaaaa, 0xbbbbbbbb]);
+    let name = &fips_version.name[..];
+    assert_eq!(name, FipsVersionCmd::NAME.as_bytes());
+}
+
+#[test]
+fn test_fips_shutdown() {
+    let mut model = run_rt_test(None, None, None);
+
+    model.step_until(|m| m.soc_mbox().status().read().mbox_fsm_ps().mbox_idle());
+
+    // SHUTDOWN
+    let payload = MailboxReqHeader {
+        chksum: caliptra_common::checksum::calc_checksum(u32::from(CommandId::SHUTDOWN), &[]),
+    };
+
+    let resp = model.mailbox_execute(u32::from(CommandId::SHUTDOWN), payload.as_bytes());
+    assert!(resp.is_ok());
+
+    // Check we are rejecting additional commands with the shutdown error code.
+    let expected_err = Err(ModelError::MailboxCmdFailed(u32::from(
+        CaliptraError::RUNTIME_SHUTDOWN,
+    )));
+    let payload = MailboxReqHeader {
+        chksum: caliptra_common::checksum::calc_checksum(u32::from(CommandId::VERSION), &[]),
+    };
+    let resp = model.mailbox_execute(u32::from(CommandId::VERSION), payload.as_bytes());
+    assert_eq!(resp, expected_err);
+}
